pyproject.toml
[tool.poetry]
name = "llm-chess"
version = "0.1.0"
description = ""
authors = ["Mathew Wang"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-chess = "^1.999"
stockfish = "^3.28.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
